fix: is_admin not being correctly validated
Due to it being 'optional' the POST /signup route would always display the Error when there was the 'matches' rule as 'true'.

After changing it to 'is_bool' and made the Controller check it beforehand and then assign ['is_admin'] the correct boolean value based on if it already exists and there's a true string assigned.